requirements.txt
streamlit==1.31.0
langchain==0.1.14
langchain-community>=0.0.17,<0.1
langchain-google-genai==0.0.5
google-generativeai==0.3.2
faiss-cpu==1.7.4
pypdf2==3.0.1
pandas==2.1.4
python-dotenv==1.0.0
matplotlib==3.8.2
crewai==0.30.11
pydantic==2.7.0
